feat: cache for allLaunches list
